Change AdBlockRule to not be a QObject (Like QWidget, QComboBox, QIODevice), but to be a container like QUrl, QString, and QImage.
AdBlockRule does not emit any signals, have any slots or properties.

Cleanup many aspects of the AdBlockRule class and fix it so it can
pass all of the existing adblockrule autotests.

Remove NOT_IMPLEMENTED_ADBLOCK_FEATURES from the AdBlockRule
autotest.  AdBlockRule must be able to parse all of the AbBlock
rules even if they are not supported.  When they are not
supported match must return false for those features.